+++ excalidraw-app/pens/PenToolbar/PenToolbar.tsx
@@ -282,4 +282,3 @@ export const PenToolbar: React.FC<PenToolbarProps> = ({ excalidrawAPI }) => {
 };
 
 export default PenToolbar;
-